22F1000598 2/Root/MrBean_Grocery_App/backend/requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
bcrypt==4.0.1
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
cachelib==0.9.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==2.3.2
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Excel==0.0.7
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.2.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
fonttools==4.46.0
html5lib==1.1
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
lml==0.1.0
MarkupSafe==2.1.3
numpy==1.26.2
packaging==23.2
pandas==2.1.4
passlib==1.7.4
Pillow==10.1.0
plotly==5.18.0
prompt-toolkit==3.0.41
pycparser==2.21
pydyf==0.8.0
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyphen==0.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
tenacity==8.2.3
texttable==1.7.0
tinycss2==1.2.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
weasyprint==60.2
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.1
zipp==3.17.0
zopfli==0.2.3
kaleido==0.2.1

